leave one cell in quiltView
sometimes ,there is only one cell in quiltView ,at this time,*top == *bottom,if we remove the only cell ,then we can not see any cell.
I often use class with "top view" or "bottom view", yeah! for example ,I set the margin top  300dp,and I add a TopView in the quiltView ,the TopView like the head of UITableView, but When I drag it down until the first cell disappear in quiltView ,I loosen. I am so puzzled that I can only see the TopView , the cell disappear,but when I swipe the quiltView fastly ,sometimes all cells has appear.Although like this，I think the quiltView still has one flaw，so I restore it .